requirements.txt
Flask==2.3.3
supabase==2.3.0
python-dotenv==1.0.0
pytz==2023.3
gunicorn==21.2.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
httpx==0.24.1
gotrue==2.9.1